rename files with incorrect capitalization
