+++ app/controllers/shortcodeController.ts
