+++ app/models/shortcodeStatistics.ts
