+++ app/services/generatorService.ts
