+++ app/services/shortcodeService.ts
@@ -1,14 +1,14 @@
-import {IShortcodeStorage} from "../storage/ishortcodestorage";
-import {ShortcodeGenerator} from "./shortcodeGeneratorService";
+import {IStorage} from "../storage/iStorage";
+import {ShortcodeGenerator} from "./generatorService";
 import {ShortcodeExistsError} from "../errors/shortcodeExistsError";
 import {ShortcodeStatistics} from "../models/shortcodeStatistics";
 export class ShortcodeService {
-    constructor(private shortcodeStorage: IShortcodeStorage, private shortcodeGenerator: ShortcodeGenerator){
+    constructor(private storage: IStorage, private shortcodeGenerator: ShortcodeGenerator){
 
     }
 
     get(shortcode: string): string {
-        return this.shortcodeStorage.get(shortcode);
+        return this.storage.get(shortcode);
     }
 
     save(url: string, shortcode?: string){
@@ -20,25 +20,25 @@ export class ShortcodeService {
            shortcode = this.generateShortCode();
         }
 
-        this.shortcodeStorage.add(shortcode, url);
-        this.shortcodeStorage.saveStatistics(shortcode, new ShortcodeStatistics());
+        this.storage.add(shortcode, url);
+        this.storage.saveStatistics(shortcode, new ShortcodeStatistics());
         return shortcode;
     }
-    
+
     updateUsage(shortcode: string){
-        let statistics = this.shortcodeStorage.getStatistics(shortcode);
+        let statistics = this.storage.getStatistics(shortcode);
         statistics.update();
-        this.shortcodeStorage.saveStatistics(shortcode, statistics);
+        this.storage.saveStatistics(shortcode, statistics);
     }
 
     exists(shortcode: string): boolean {
         return !!this.get(shortcode);
     }
-    
+
     getStatistics(shortcode: string): ShortcodeStatistics {
-        return this.shortcodeStorage.getStatistics(shortcode);
+        return this.storage.getStatistics(shortcode);
     }
-    
+
     private generateShortCode(): string{
         let shortcode;
         do{
