+++ gulpfile.js
@@ -36,7 +36,7 @@ var compileTs = function (sourceTsFiles, srcOptions, destination, tsProject) {
         .pipe(gulp.dest(destination));
 };
 
-gulp.task('compile-e2e', ['compile-all'], function(){
+gulp.task('compile-e2e', function(){
     return compileTs([config.allE2ETests], {base: './e2e'}, config.compiledE2EPath, tsProjectE2E);
 });
